Rename files: rams.js/main.css to index.js/index.css
