+++ README.md
@@ -43,12 +43,12 @@ Rams is a class-less css design system framework that utilizes the power of data
 
 Edit the [variables](https://github.com/jacobxperez/rams/blob/master/src/css/var.css), add
 [custom](https://github.com/jacobxperez/rams/tree/master/src/css/custom) modules and linked
-them to [main.css](https://github.com/jacobxperez/rams/blob/master/src/css/main.css) for compiling.
+them to [index.css](https://github.com/jacobxperez/rams/blob/master/src/css/index.css) for compiling.
 
 - [src/css](https://github.com/jacobxperez/rams/tree/master/src/css/)
   - [core](https://github.com/jacobxperez/rams/tree/master/src/css/core) (important files for compiling)
   - [custom](https://github.com/jacobxperez/rams/tree/master/src/css/custom) (add custom modules here)
-  - [main.css](https://github.com/jacobxperez/rams/blob/master/src/css/main.css) (compile to produce the final CSS)
+  - [index.css](https://github.com/jacobxperez/rams/blob/master/src/css/index.css) (compile to produce the final CSS)
   - [var.css](https://github.com/jacobxperez/rams/blob/master/src/css/var.css) (variables for font, colors, and more)
 
 **note**: When starting a project with Rams we recommend **not** to edit the [core](https://github.com/jacobxperez/rams/tree/master/src/css/core) files unless you are [contributing](https://github.com/jacobxperez/rams/blob/master/.github/CONTRIBUTING.md) code back to Rams. Any custom modules should go in the [custom](https://github.com/jacobxperez/rams/tree/master/src/css/custom) folder, this will help you separate your code from Rams.
