+++ src/css/custom/README.md
@@ -1,4 +1,4 @@
 # Custom Modules
 
 - Modify print styles to your project specifications.
-- Add custom modules here then link them on file [main.css](https://github.com/jacobxperez/rams/blob/master/src/css/main.css) for compiling.
+- Add custom modules here then link them on file [index.css](https://github.com/jacobxperez/rams/blob/master/src/css/index.css) for compiling.
