+++ src/docs/css/index.css
@@ -6,7 +6,7 @@
  */
 
 /* Main from src */
-@import "../../../src/css/main.css";
+@import "../../../src/css/index.css";
 
 /* Custom Modules */
 @import "custom/sections.css";
