+++ src/docs/icons/index.html
@@ -8,8 +8,8 @@
         <meta name="author" content="Jacob Perez" />
         <title>Rams - Icons</title>
         <link rel="icon" href="../favicon.svg" />
-        <link rel="preload" href="../templates/main.html" as="fetch" type="text/html" crossorigin />
-        <link rel="stylesheet" href="../css/main.css" />
+        <link rel="preload" href="../templates/index.html" as="fetch" type="text/html" crossorigin />
+        <link rel="stylesheet" href="../css/index.css" />
         <script src="../js/index.js"></script>
         <script type="module" src="../js/defer.js"></script>
         <script>
