+++ src/docs/index.html
@@ -15,8 +15,8 @@
         />
         <title>Rams - Class-Less CSS Design System Framework</title>
         <link rel="icon" href="./favicon.svg" />
-        <link rel="preload" href="templates/main.html" as="fetch" type="text/html" crossorigin />
-        <link rel="stylesheet" href="css/main.css" />
+        <link rel="preload" href="templates/index.html" as="fetch" type="text/html" crossorigin />
+        <link rel="stylesheet" href="css/index.css" />
         <script src="./js/index.js"></script>
         <script type="module" src="./js/defer.js"></script>
         <script>
@@ -70,8 +70,8 @@
                 >, add
                 <a href="https://github.com/jacobxperez/rams/tree/master/src/css/custom">custom</a>
                 modules and linked them to
-                <a href="https://github.com/jacobxperez/rams/blob/master/src/css/main.css"
-                    >main.css</a
+                <a href="https://github.com/jacobxperez/rams/blob/master/src/css/index.css"
+                    >index.css</a
                 >
                 for compiling.
             </p>
@@ -94,8 +94,8 @@
                         </li>
                         <li>
                             <a
-                                href="https://github.com/jacobxperez/rams/blob/master/src/css/main.css"
-                                >main.css</a
+                                href="https://github.com/jacobxperez/rams/blob/master/src/css/index.css"
+                                >index.css</a
                             >
                             (compile to produce the final CSS)
                         </li>
