+++ src/index.html
@@ -7,7 +7,7 @@
         <meta name="description" content="" />
         <meta name="author" content="" />
         <title>Read Me</title>
-        <link rel="stylesheet" href="css/main.css" />
+        <link rel="stylesheet" href="css/index.css" />
         <script type="module" src="js/index.js" defer></script>
     </head>
 
@@ -105,8 +105,8 @@
                         >custom</a
                     >
                     modules and linked them to
-                    <a href="https://github.com/jacobxperez/rams/blob/master/src/css/main.css"
-                        >main.css</a
+                    <a href="https://github.com/jacobxperez/rams/blob/master/src/css/index.css"
+                        >index.css</a
                     >
                     for compiling.
                 </p>
@@ -131,8 +131,8 @@
                             </li>
                             <li>
                                 <a
-                                    href="https://github.com/jacobxperez/rams/blob/master/src/css/main.css"
-                                    >main.css</a
+                                    href="https://github.com/jacobxperez/rams/blob/master/src/css/index.css"
+                                    >index.css</a
                                 >
                                 (compile to produce the final CSS)
                             </li>
